requirements.txt
streamlit
requests
requests-oauthlib
python-dotenv